Add Storybook Center Decorator
