+++ src/app/components/elements/Author/Author.story.jsx
@@ -5,18 +5,17 @@ import rootReducer from 'app/redux/RootReducer';
 import { Provider } from 'react-redux';
 import { createStore } from 'redux';
 import Author from './index';
-import { Center } from '../Tooltip.story';
+import { Center } from '../../../../../.storybook/decorators';
 import { IntlProvider } from 'react-intl';
 
 const store = createStore(rootReducer);
 
 storiesOf('Elements', module)
     .addDecorator(withKnobs)
+    .addDecorator(Center)
     .addDecorator(getStory => <Provider store={store}>{getStory()}</Provider>)
     .add('Author', () => (
         <IntlProvider locale="en">
-            <Center>
-                <Author author={'maitland'} />
-            </Center>
+            <Author author={'maitland'} />
         </IntlProvider>
     ));
