+++ src/app/components/elements/ChangePassword.story.jsx
@@ -5,20 +5,19 @@ import rootReducer from 'app/redux/RootReducer';
 import { Provider } from 'react-redux';
 import { createStore } from 'redux';
 import ChangePassword from './ChangePassword';
-import { Center } from './Tooltip.story';
+import { Center } from '../../../../.storybook/decorators';
 
 const store = createStore(rootReducer);
 
 storiesOf('Elements', module)
     .addDecorator(withKnobs)
+    .addDecorator(Center)
     .addDecorator(getStory => <Provider store={store}>{getStory()}</Provider>)
     .add('ChangePassword', () => (
-        <Center>
-            <ChangePassword
-                username={'maitland'}
-                defaultPassword={'password1'}
-                authType={null}
-                priorAuthKey={'posting'} // Required pubkey if authType is given
-            />
-        </Center>
+        <ChangePassword
+            username={'maitland'}
+            defaultPassword={'password1'}
+            authType={null}
+            priorAuthKey={'posting'} // Required pubkey if authType is given
+        />
     ));
