+++ src/app/components/elements/Follow/Follow.story.jsx
@@ -5,22 +5,21 @@ import rootReducer from 'app/redux/RootReducer';
 import { Provider } from 'react-redux';
 import { createStore } from 'redux';
 import Follow from './index';
-import { Center } from '../Tooltip.story';
+import { Center } from '../../../../../.storybook/decorators';
 
 const store = createStore(rootReducer);
 
 storiesOf('Elements', module)
     .addDecorator(withKnobs)
+    .addDecorator(Center)
     .addDecorator(getStory => <Provider store={store}>{getStory()}</Provider>)
     .add('Follow', () => (
-        <Center>
-            <Follow
-                className="float-right"
-                follower={'maitland'}
-                following={'Carol'}
-                what="blog"
-                showFollow={boolean('Show Follow?', true)}
-                showMute={boolean('Show Mute?', true)}
-            />
-        </Center>
+        <Follow
+            className="float-right"
+            follower={'maitland'}
+            following={'Carol'}
+            what="blog"
+            showFollow={boolean('Show Follow?', true)}
+            showMute={boolean('Show Mute?', true)}
+        />
     ));
