+++ src/app/components/elements/PasswordInput.story.jsx
@@ -1,24 +1,26 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import PasswordInput from './PasswordInput';
-import { Center } from './Tooltip.story';
+import { Center } from '../../../../.storybook/decorators';
 
-storiesOf('Elements', module).add('PasswordInput', () => (
-    <Center>
-        <PasswordInput
-            passwordLabel={'Enter New Password'}
-            onChange={() => {}}
-        />
-        <PasswordInput
-            passwordLabel={'Confirm Password'}
-            inputConfirmPassword={true}
-            onChange={() => {}}
-        />
-        <PasswordInput
-            passwordLabel={'Disabled Password'}
-            inputConfirmPassword={false}
-            disabled={true}
-            onChange={() => {}}
-        />
-    </Center>
-));
+storiesOf('Elements', module)
+    .addDecorator(Center)
+    .add('PasswordInput', () => (
+        <div>
+            <PasswordInput
+                passwordLabel={'Enter New Password'}
+                onChange={() => {}}
+            />
+            <PasswordInput
+                passwordLabel={'Confirm Password'}
+                inputConfirmPassword={true}
+                onChange={() => {}}
+            />
+            <PasswordInput
+                passwordLabel={'Disabled Password'}
+                inputConfirmPassword={false}
+                disabled={true}
+                onChange={() => {}}
+            />
+        </div>
+    ));
