+++ src/app/components/elements/Reputation.story.jsx
@@ -1,10 +1,8 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import Reputation from './Reputation';
-import { Center } from './Tooltip.story';
+import { Center } from '../../../../.storybook/decorators';
 
-storiesOf('Elements', module).add('Reputation', () => (
-    <Center>
-        <Reputation value={1200} />
-    </Center>
-));
+storiesOf('Elements', module)
+    .addDecorator(Center)
+    .add('Reputation', () => <Reputation value={1200} />);
