+++ src/app/components/elements/TimeAgoWrapper/story.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs, date } from '@storybook/addon-knobs';
+import { Center } from '../../../../../.storybook/decorators';
 
 import { IntlProvider } from 'react-intl';
 
@@ -8,6 +9,7 @@ import TimeAgoWrapper from './';
 
 storiesOf('Elements', module)
     .addDecorator(withKnobs)
+    .addDecorator(Center)
     .add('TimeAgoWrapper', () => (
         <IntlProvider locale="en">
             <TimeAgoWrapper date={date('date', new Date('1 Jul 2016'))} />
