+++ src/app/components/elements/TransactionError.story.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs, text } from '@storybook/addon-knobs';
 import TransactionError from './TransactionError';
-import { Center } from './Tooltip.story';
+import { Center } from '../../../../.storybook/decorators';
 import rootReducer from 'app/redux/RootReducer';
 import { Provider } from 'react-redux';
 import { createStore } from 'redux';
@@ -10,12 +10,11 @@ const store = createStore(rootReducer);
 
 storiesOf('Elements', module)
     .addDecorator(withKnobs)
+    .addDecorator(Center)
     .addDecorator(getStory => <Provider store={store}>{getStory()}</Provider>)
     .add('TransactionError', () => (
-        <Center>
-            <TransactionError
-                opType="This is a test"
-                error="There was an error ey"
-            />
-        </Center>
+        <TransactionError
+            opType="This is a test"
+            error="There was an error ey"
+        />
     ));
