+++ src/app/components/elements/Userpic.story.jsx
@@ -5,16 +5,13 @@ import rootReducer from 'app/redux/RootReducer';
 import { Provider } from 'react-redux';
 import { createStore } from 'redux';
 import Userpic from './Userpic';
-import { Center } from './Tooltip.story';
+import { Center } from '../../../../.storybook/decorators';
 
 const store = createStore(rootReducer);
 global.$STM_Config = { img_proxy_prefix: 'https://steemitimages.com/' };
 
 storiesOf('Elements', module)
     .addDecorator(withKnobs)
+    .addDecorator(Center)
     .addDecorator(getStory => <Provider store={store}>{getStory()}</Provider>)
-    .add('Userpic', () => (
-        <Center>
-            <Userpic account={'maitland'} />
-        </Center>
-    ));
+    .add('Userpic', () => <Userpic account={'maitland'} />);
