+++ src/app/components/elements/WalletSubMenu.story.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs, text } from '@storybook/addon-knobs';
 import WalletSubMenu from './WalletSubMenu';
-import { Center } from './Tooltip.story';
+import { Center } from '../../../../.storybook/decorators';
 import rootReducer from 'app/redux/RootReducer';
 import { Provider } from 'react-redux';
 import { createStore } from 'redux';
@@ -10,9 +10,6 @@ const store = createStore(rootReducer);
 
 storiesOf('Elements', module)
     .addDecorator(withKnobs)
+    .addDecorator(Center)
     .addDecorator(getStory => <Provider store={store}>{getStory()}</Provider>)
-    .add('WalletSubMenu', () => (
-        <Center>
-            <WalletSubMenu />
-        </Center>
-    ));
+    .add('WalletSubMenu', () => <WalletSubMenu />);
